Revert "Støtt flyttall sum til utbetaling" som en del av å flytte endringer til branch
This reverts commit ea57ce03c6b24feea78ac88a63cb923507a81238.

+++ src/main/kotlin/no/nav/hjelpemidler/brille/utbetaling/Utbetaling.kt
@@ -23,7 +23,7 @@ data class Utbetaling(
         klassekode = "BARNEBRILLER",
         fom = LocalDate.now(),
         tom = LocalDate.now(),
-        sats = vedtak.beløp,
+        sats = vedtak.beløp.toInt(),
         satstype = "ENG",
         saksbehandler = "BB",
     )
